backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

deepface==0.0.93
tensorflow==2.16.1
torch
torchvision


sqlalchemy==2.0.36

python-multipart==0.0.12
python-dotenv==1.0.1

Pillow==10.4.0
requests==2.32.3

gunicorn
locust==2.32.3

python-jose[cryptography]==3.3.0
bcrypt==4.0.1